Luigi/requirements.txt
connected-components-3d==3.26.1
dataclasses==0.6
et_xmlfile==2.0.0
ImageIO==2.37.2
joblib==1.5.2
lazy_loader==0.4
lockfile==0.12.2
luigi==3.6.0
networkx==3.6.1
nibabel==5.3.3
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psutil==7.1.3
pydicom==3.0.1
pynrrd==1.1.3
pysera==2.1.5
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
rt-utils==1.2.7
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
shiboken6==6.10.1
simpleitk==2.5.3
six==1.17.0
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
